requirements.txt
anthropic>=0.49.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
playwright>=1.44.0
